+++ source/extensions/omni.isaac.conveyor/python/scripts/extension.py
@@ -17,9 +17,16 @@ import weakref
 from ..bindings._omni_isaac_conveyor import acquire_interface as _acquire
 from ..bindings._omni_isaac_conveyor import release_interface as _release
 
+from .style import UI_STYLES
+
+from .ui.conveyor_builder import ConveyorBuilderWidget
+
+EXTENSION_NAME = "Conveyor Utility"
+
 
 class Extension(omni.ext.IExt):
     def __init__(self) -> None:
+        self.widget = None
         menu_items = [
             MenuItemDescription(
                 name="Warehouse Items",
@@ -28,16 +35,42 @@ class Extension(omni.ext.IExt):
                 ],
             )
         ]
+        self._menu_items_2 = [
+            MenuItemDescription(name="Conveyor Track Builder", onclick_fn=lambda a=weakref.proxy(self): a.create_ui())
+        ]
 
         self._menu_items = [MenuItemDescription(name="Isaac", glyph="plug.svg", sub_menu=menu_items)]
 
         add_menu_items(self._menu_items, "Create")
+        add_menu_items(self._menu_items_2, "Tools")
 
-    def on_startup(self):
+    def on_startup(self, ext_id: str):
         self.__interface = _acquire()
+        theme = "NvidiaDark"
+        self.ext_id = ext_id
+        ext_path = omni.kit.app.get_app().get_extension_manager().get_extension_path(ext_id)
+        self.ext_path = ext_path + "/omni/isaac/conveyor"
+        self.config_file = ext_path + "/data/track_types.json"
+        self.mdl_file = ext_path + "/data/GhostVolumetric.mdl"
+        self._style = UI_STYLES[theme]
+        for i in [a for a in self._style.keys() if "{}" in self._style[a].get("image_url", "")]:
+            self._style[i]["image_url"] = self._style[i]["image_url"].format(self.ext_path)
+
+    def create_ui(self):
+        self._window = ui.Window("Conveyor Builder", open=True, dock=ui.DockPreference.LEFT_BOTTOM)
+        self._window.set_visibility_changed_fn(self.on_visibility_changed)
+        with self._window.frame:
+            self.widget = ConveyorBuilderWidget(self.config_file, mdl_file=self.mdl_file, style=self._style)
+
+    def on_visibility_changed(self, value):
+        if not value:
+            self.widget.shutdown()
 
     def on_shutdown(self):
         remove_menu_items(self._menu_items, "Create")
+        remove_menu_items(self._menu_items_2, "Tools")
+        if self.widget:
+            self.widget.shutdown()
         _release(self.__interface)
         self.__interface = None
         gc.collect()